sound_detector_test_app/requirements.txt
flask
flask-sock
ndi-python
ffmpeg-python